// packages/backend/src/shared/common/domain/enums/autonomous-communities.enum.ts
/** Autonomous communities of Spain */
export enum AutonomousCommunities {
	ANDALUCIA = 'ANDALUCIA',
	ARAGON = 'ARAGON',
	ASTURIAS = 'ASTURIAS',
	BALEARES = 'BALEARES',
	CANARIAS = 'CANARIAS',
	CANTABRIA = 'CANTABRIA',
	CASTILLA_LA_MANCHA = 'CASTILLA_LA_MANCHA',
	CASTILLA_LEON = 'CASTILLA_LEON',
	CATALUNA = 'CATALUNA',
	COMUNIDAD_VALENCIANA = 'COMUNIDAD_VALENCIANA',
	EXTREMADURA = 'EXTREMADURA',
	GALICIA = 'GALICIA',
	LA_RIOJA = 'LA_RIOJA',
	MADRID = 'MADRID',
	MURCIA = 'MURCIA',
	NAVARRA = 'NAVARRA',
	PAIS_VASCO = 'PAIS_VASCO',
	CEUTA = 'CEUTA',
	MELILLA = 'MELILLA',
}