// packages/backend/src/shared/common/domain/enums/provinces.enum.ts
/**
 * All provinces of Spain
 */
export enum Provinces {
	A_CORUNA = 'A_CORUNA',
	ALAVA = 'ALAVA',
	ALBACETE = 'ALBACETE',
	ALICANTE = 'ALICANTE',
	ALMERIA = 'ALMERIA',
	ASTURIAS = 'ASTURIAS',
	AVILA = 'AVILA',
	BADAJOZ = 'BADAJOZ',
	BALEARES = 'BALEARES',
	BARCELONA = 'BARCELONA',
	BURGOS = 'BURGOS',
	CACERES = 'CACERES',
	CADIZ = 'CADIZ',
	CANTABRIA = 'CANTABRIA',
	CASTELLON = 'CASTELLON',
	CEUTA = 'CEUTA',
	CIUDAD_REAL = 'CIUDAD_REAL',
	CORDOBA = 'CORDOBA',
	CUENCA = 'CUENCA',
	GIRONA = 'GIRONA',
	GRANADA = 'GRANADA',
	GUADALAJARA = 'GUADALAJARA',
	GIPUZKOA = 'GIPUZKOA',
	HUELVA = 'HUELVA',
	HUESCA = 'HUESCA',
	JAEN = 'JAEN',
	LA_RIOJA = 'LA_RIOJA',
	LAS_PALMAS = 'LAS_PALMAS',
	LEON = 'LEON',
	LERIDA = 'LERIDA',
	LUGO = 'LUGO',
	MADRID = 'MADRID',
	MALAGA = 'MALAGA',
	MELILLA = 'MELILLA',
	MURCIA = 'MURCIA',
	NAVARRA = 'NAVARRA',
	OURENSE = 'OURENSE',
	PALENCIA = 'PALENCIA',
	PONTEVEDRA = 'PONTEVEDRA',
	SALAMANCA = 'SALAMANCA',
	SEGOVIA = 'SEGOVIA',
	SEVILLA = 'SEVILLA',
	SORIA = 'SORIA',
	TARRAGONA = 'TARRAGONA',
	TENERIFE = 'TENERIFE',
	TERUEL = 'TERUEL',
	TOLEDO = 'TOLEDO',
	VALENCIA = 'VALENCIA',
	VALLADOLID = 'VALLADOLID',
	VIZCAYA = 'VIZCAYA',
	ZAMORA = 'ZAMORA',
	ZARAGOZA = 'ZARAGOZA',
}